refactor: empaquetar laberintos sin diagonales
