--- src/SolverEjercicio3A.java
@@ -1,17 +1,0 @@
-import Tarea1.SolverEjercicio1A;
-
-public class SolverEjercicio3A extends SolverEjercicio1A {
-//    @Override
-//    public int[][] getMovimientos() {
-//        return new int[][] {
-//                {-1, 0},
-//                {-1, 1},
-//                {0, 1},
-//                {1, 1},
-//                {1, 0},
-//                {1, -1},
-//                {0, -1},
-//                {-1, -1}
-//        };
-//    }
-}
